requirements.txt
Django==4.2.3
django-taggit==4.0.0
django-tailwind==3.6.0
django_browser_reload==1.10.0
django-modeltranslation == 0.18.11
requests==2.31.0
django-pwa==1.1.0